B1 - move PGPContentSignerBuilderProviderFactory into openpgp/api/operator
