+++ pg/src/main/java/org/bouncycastle/openpgp/api/operator/PGPContentSignerBuilderProviderFactory.java
@@ -1,8 +1,9 @@
-package org.bouncycastle.openpgp.operator;
+package org.bouncycastle.openpgp.api.operator;
 
 import org.bouncycastle.openpgp.PGPException;
 import org.bouncycastle.openpgp.api.KeyPassphraseProvider;
 import org.bouncycastle.openpgp.api.OpenPGPKey;
+import org.bouncycastle.openpgp.operator.PGPContentSignerBuilderProvider;
 
 public interface PGPContentSignerBuilderProviderFactory
 {
