+++ pgsc/src/main/java/org/bouncycastle/openpgp/smartcard/OpenPGPSmartCardManager.java
@@ -5,7 +5,7 @@ import org.bouncycastle.openpgp.api.KeyPassphraseProvider;
 import org.bouncycastle.openpgp.api.OpenPGPKey;
 import org.bouncycastle.openpgp.api.PublicKeyDataDecryptorFactoryProvider;
 import org.bouncycastle.openpgp.operator.PGPContentSignerBuilderProvider;
-import org.bouncycastle.openpgp.operator.PGPContentSignerBuilderProviderFactory;
+import org.bouncycastle.openpgp.api.operator.PGPContentSignerBuilderProviderFactory;
 import org.bouncycastle.openpgp.operator.PublicKeyDataDecryptorFactory;
 import org.bouncycastle.openpgp.smartcard.card.CardException;
